// src/api/orval/model/registerDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * COOP API
 * API completa para gerenciamento de projetos colaborativos com autenticação JWT e recuperação de senha via Magic Link
 * OpenAPI spec version: 1.0
 */

export interface RegisterDto {
  /**
   * Nome de usuário único
   * @minLength 3
   * @maxLength 50
   */
  username: string;
  /** Email do usuário */
  email: string;
  /**
   * Senha do usuário
   * @minLength 6
   */
  password: string;
}